Various improvements
- Optimizes certain parts of code
- Fixed an issue where the store could have the game crash
- Fixed an issue where the use could type in negative numbers at the store

+++ Oregon_Trail.cpp
@@ -331,7 +331,7 @@ void Oregon_Trail::AddToLeaderBoard() {
 	gameScore.score = m_trailGame.GetTotalScore();
 
 	// Cycle through leaderboard scores and if the player's is higher, insert it into the leaderboard
-	for (int i = 0; i < m_entries.size(); i++) {
+	for (size_t i = 0; i < m_entries.size(); i++) {
 		if (m_trailGame.GetTotalScore() > m_entries[i].score) {
 
 			cout << "\t Enter your name: ";
